+++ TopModel.Generator.Jpa/AssociationType.cs
@@ -1,0 +1,8 @@
+namespace TopModel.Generator.Jpa;
+
+public enum AssociationType
+{
+    OneToOne,
+    OneToMany,
+    ManyToOne,
+}