+++ TopModel.Generator.Jpa/CascadeType.cs
@@ -1,0 +1,12 @@
+namespace TopModel.Generator.Jpa;
+
+public enum CascadeType
+{
+    All,
+    Persist,
+    Merge,
+    Remove,
+    Refresh,
+    Detach,
+    Lock,
+}